[feat] add pcss [fix] fix sss jagged-ness

+++ shaders/include/utility/vogel_disk_blur.glsl
@@ -1,21 +1,21 @@
 #if !defined INCLUDE_VOGEL_DISK_BLUR
-#define INCLUDE_VOGEL_DISK_BLUR
+    #define INCLUDE_VOGEL_DISK_BLUR
 
-#include "/include/utility/dither.glsl"
+    #include "/include/utility/dither.glsl"
 
-#define GOLDEN_ANGLE 2.39996322972865332 // Radians.
+    #define GOLDEN_ANGLE 2.39996322972865332 // Radians.
 
-float _compute_vogel_disk_radius(float idx, int sample_count);
+    float _compute_vogel_disk_radius(uint idx, int sample_count);
 
-vec2 compute_vogel_disk_sample_uv(float idx, int sample_count) {
-    float dither = compute_dither(gl_FragCoord.xy);
-    float radius = _compute_vogel_disk_radius(idx, sample_count);
-    float theta = (idx + 1.0) * GOLDEN_ANGLE;
+    vec2 compute_vogel_disk_sample_uv(uint idx, int sample_count) {
+        float dither = compute_dither(gl_FragCoord.xy);
+        float radius = _compute_vogel_disk_radius(idx, sample_count);
+        float theta = float(idx + 1.0) * GOLDEN_ANGLE;
 
-    return radius * vec2(cos(theta), sin(theta));
-}
+        return radius * vec2(cos(theta), sin(theta));
+    }
 
-float _compute_vogel_disk_radius(float idx, int sample_count) {
-    return sqrt((idx + 0.5) / sample_count);
-}
+    float _compute_vogel_disk_radius(uint idx, int sample_count) {
+        return sqrt((float(idx) + 0.5) / sample_count);
+    }
 #endif
